refactor: Merge edit and spatial skeleton errors
