+++ src/layer/segmentation/index.ts
@@ -57,8 +57,8 @@ import {
   executeSpatialSkeletonNodeRadiusUpdate,
   executeSpatialSkeletonReroot,
   executeSpatialSkeletonNodeTrueEndUpdate,
+  showSpatialSkeletonActionError,
 } from "#src/layer/segmentation/spatial_skeleton_commands.js";
-import { showSpatialSkeletonActionError } from "#src/layer/segmentation/spatial_skeleton_errors.js";
 import {
   MeshLayer,
   MeshSource,
